requirements.txt
# Критичные зависимости для Disco Coop
python-telegram-bot>=22.0,<23.0
aiohttp>=3.8.0,<4.0.0
pyyaml>=6.0
loguru>=0.7.0
Pillow>=9.0.0
requests>=2.28.0

# Дополнительные зависимости (опциональные)
opencv-python-headless>=4.5.0
numpy>=1.20.0
six>=1.16.0
easyocr>=1.7.0

# Инструменты сборки (устанавливаются автоматически скриптом)
setuptools>=65.0.0
wheel>=0.37.0
build>=0.8.0

# Игровые контроллеры (устанавливаются отдельно в install.sh с обработкой ошибок) 
# PyAutoGUI>=0.9.50  # Требует системные библиотеки X11 на Linux
# pynput>=1.7.0      # Может требовать evdev на Linux

# LLM integration
# Ollama устанавливается отдельно через install.sh скрипт